Update: Reworked FCT class
