+++ src/batchMultiSigCall/classes/FCTCalls/index.ts
@@ -1,0 +1,2 @@
+export * from "./FCTCalls";
+export * from "./helpers";
