+++ src/batchMultiSigCall/classes/index.ts
@@ -1,3 +1,4 @@
 export * from "./EIP712";
 export * from "./ExportFCT";
+export * from "./FCTCalls";
 export * from "./Options";
