+++ src/batchMultiSigCall/helpers/index.ts
@@ -1,4 +1,3 @@
-export * from "./checkers";
 export * from "./eip712";
 export * from "./handlers";
 export * from "./sessionId";
