+++ src/batchMultiSigCall/methods/index.ts
@@ -1,6 +1,5 @@
 export * from "./checkers";
 export * from "./FCT";
-export * from "./getters";
 export * from "./helpers";
 export * from "./plugins";
 export * from "./variables";
